CS50 Cash
Harvard's CS50 Class - Problem:  Cash

For this problem, we had to create a program in C, that correctly calculated how many coins were needed, to create the correct change.  The user would have to input the change amount, and the program would provide how many coins the user received in return.